# specs/openapi.yaml
openapi: 3.1.0
info:
  title: FlowProof API
  version: 0.6.0
paths:
...
  /api/v1/recovery-plans/{plan_id}:
    get:
      tags:
      - recovery
      summary: Recovery Plan
      operationId: recovery_plan_api_v1_recovery_plans__plan_id__get
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          title: Plan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Recovery Plan Api V1 Recovery Plans  Plan Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/recovery-plans/{plan_id}/approve:
    post:
      tags:
      - recovery
      summary: Approve Recovery
      operationId: approve_recovery_api_v1_recovery_plans__plan_id__approve_post
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          title: Plan Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Approve Recovery Api V1 Recovery Plans  Plan Id  Approve
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/recovery-plans/{plan_id}/reject:
    post:
      tags:
      - recovery
      summary: Reject Recovery
      operationId: reject_recovery_api_v1_recovery_plans__plan_id__reject_post
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          title: Plan Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryDecisionIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reject Recovery Api V1 Recovery Plans  Plan Id  Reject
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/recovery-plans/{plan_id}/revoke-approval:
    post:
      tags:
      - recovery
      summary: Revoke Recovery Approval
      operationId: revoke_recovery_approval_api_v1_recovery_plans__plan_id__revoke_approval_post
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          title: Plan Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryDecisionIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Revoke Recovery Approval Api V1 Recovery Plans  Plan
                  Id  Revoke Approval Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/recovery-plans/{plan_id}/reconcile:
    post:
      tags:
      - recovery
      summary: Reconcile Recovery
      operationId: reconcile_recovery_api_v1_recovery_plans__plan_id__reconcile_post
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          title: Plan Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReconcileIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reconcile Recovery Api V1 Recovery Plans  Plan Id  Reconcile
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/recovery-plans/{plan_id}/attempts:
    get:
      tags:
      - recovery
      summary: Recovery Attempts
      operationId: recovery_attempts_api_v1_recovery_plans__plan_id__attempts_get
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          title: Plan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Recovery Attempts Api V1 Recovery Plans  Plan Id  Attempts
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/recovery-plans/{plan_id}/decisions:
    get:
      tags:
      - recovery
      summary: Recovery Decisions
      operationId: recovery_decisions_api_v1_recovery_plans__plan_id__decisions_get
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          title: Plan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Recovery Decisions Api V1 Recovery Plans  Plan Id  Decisions
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/recovery-plans/{plan_id}/operator-review:
    post:
      tags:
      - recovery
      summary: Record Operator Review
      operationId: record_operator_review_api_v1_recovery_plans__plan_id__operator_review_post
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          title: Plan Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorReviewIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Record Operator Review Api V1 Recovery Plans  Plan
                  Id  Operator Review Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/recovery-plans/{plan_id}/execute:
    post:
      tags:
      - recovery
      summary: Execute Recovery
      operationId: execute_recovery_api_v1_recovery_plans__plan_id__execute_post
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          title: Plan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Execute Recovery Api V1 Recovery Plans  Plan Id  Execute
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/recovery-plans/{plan_id}/verify:
    post:
      tags:
      - recovery
      summary: Verify Recovery
      operationId: verify_recovery_api_v1_recovery_plans__plan_id__verify_post
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          title: Plan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Verify Recovery Api V1 Recovery Plans  Plan Id  Verify
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
...
components:
  schemas:
    ApprovalIn:
      properties:
        plan_hash:
          type: string
          pattern: ^[0-9a-f]{64}$
          title: Plan Hash
        incident_status:
          type: string
          pattern: ^recovery_proposed$
          title: Incident Status
        reason_code:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
          title: Reason Code
          default: approved_by_operator
        note:
          anyOf:
          - type: string
            maxLength: 500
            minLength: 1
          - type: 'null'
          title: Note
      additionalProperties: false
      type: object
      required:
      - plan_hash
      - incident_status
      title: ApprovalIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OperatorReviewIn:
      properties:
        plan_hash:
          type: string
          pattern: ^[0-9a-f]{64}$
          title: Plan Hash
        capsule_digest:
          anyOf:
          - type: string
            pattern: ^[0-9a-f]{64}$
          - type: 'null'
          title: Capsule Digest
        evidence_understood:
          type: boolean
          title: Evidence Understood
        authoritative_source_understood:
          type: boolean
          title: Authoritative Source Understood
        blast_radius_understood:
          type: boolean
          title: Blast Radius Understood
        proposed_action_understood:
          type: boolean
          title: Proposed Action Understood
        reject_path_available:
          type: boolean
          title: Reject Path Available
        revoke_path_available:
          type: boolean
          title: Revoke Path Available
        reconciliation_path_understood:
          type: boolean
          title: Reconciliation Path Understood
        final_decision:
          type: string
          pattern: ^(approve|reject|revoke_approval|observe_only)$
          title: Final Decision
        note:
          anyOf:
          - type: string
            maxLength: 500
            minLength: 1
          - type: 'null'
          title: Note
      additionalProperties: false
      type: object
      required:
      - plan_hash
      - evidence_understood
      - authoritative_source_understood
      - blast_radius_understood
      - proposed_action_understood
      - reject_path_available
      - revoke_path_available
      - reconciliation_path_understood
      - final_decision
      title: OperatorReviewIn
...
    ReconcileIn:
      properties:
        plan_hash:
          type: string
          pattern: ^[0-9a-f]{64}$
          title: Plan Hash
      additionalProperties: false
      type: object
      required:
      - plan_hash
      title: ReconcileIn
    RecoveryDecisionIn:
      properties:
        plan_hash:
          type: string
          pattern: ^[0-9a-f]{64}$
          title: Plan Hash
        incident_status:
          type: string
          pattern: ^(recovery_proposed|recovery_approved)$
          title: Incident Status
        reason_code:
          type: string
          pattern: ^[a-z][a-z0-9_]{0,63}$
          title: Reason Code
        note:
          anyOf:
          - type: string
            maxLength: 500
            minLength: 1
          - type: 'null'
          title: Note
      additionalProperties: false
      type: object
      required:
      - plan_hash
      - incident_status
      - reason_code
      title: RecoveryDecisionIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API credential issued to a human or service principal.
    SessionCookie:
      type: apiKey
      in: cookie
      name: flowproof_session
      description: Human browser session; state-changing requests also require X-CSRF-Token.